pyproject.toml
[project]
name = "rlms"
authors = [
    {name = "Alex Zhang", email = "altzhang@mit.edu"},
]
version = "0.1.0"
description = "Recursive Language Models."
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "anthropic>=0.75.0",
    "google-adk>=1.27",
    "google-genai>=1.56.0",
    "langfuse>=3.14.0",
    "openai>=2.14.0",
    "openinference-instrumentation-google-adk>=0.1.9",
    "portkey-ai>=2.1.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.24.0",
    "python-dotenv>=1.2.1",
    "repomix>=0.5.0",
    "requests>=2.32.5",
    "rich>=13.0.0",
    "substack-api>=0.1.0",
    "ipython>=8.0",
    "debugpy>=1.8",
    "pydantic>=2.12.5",
    "typer>=0.21.0",
    "google-api-python-client>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/alexzhang13/rlm"
Repository = "https://github.com/alexzhang13/rlm"
Issues = "https://github.com/alexzhang13/rlm/issues"

[project.optional-dependencies]
dashboard = ["nicegui>=2.0"]
modal = ["modal>=0.73.0", "dill>=0.3.7"]
e2b = ["e2b-code-interpreter>=0.0.11", "dill>=0.3.7"]
daytona = ["daytona>=0.128.1", "dill>=0.3.7"]
prime = ["prime-sandboxes>=0.2.0", "dill>=0.3.7"]
litellm = ["litellm>=1.50.0"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["rlm", "rlm.*", "rlm_adk", "rlm_adk.*"]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.14.10",
    "ty>=0.0.7",
]
test = [
    "aiohttp>=3.9.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests", "tests_rlm_adk"]
asyncio_mode = "auto"
addopts = '-m "provider_fake_contract and not agent_challenge"'
markers = [
    "provider_fake: e2e tests that run against provider-fake replay fixtures (no network)",
    "provider_fake_contract: default provider-fake fixture-contract suite",
    "provider_fake_extended: non-default provider-fake coverage beyond the default fixture-contract suite",
    "agent_challenge: provider-fake tests that depend on fixtures under tests_rlm_adk/fixtures/provider_fake/agent_challenge/",
    "unit_nondefault: non-default tests excluded from the provider-fake default pytest run",
    "provider_fake_litellm: provider-fake tests exercising the LiteLLM Router path (OpenAI-compatible)",
    "litellm_live: live LiteLLM integration tests requiring real API keys",
]
